STRK-116: consolidate smoke, inventory CRUD, and search specs
Consolidate 7 scattered Playwright specs into 2 compact domain suites
under tests/playwright/core/, preserving all durable assertions.

New: core/smoke.spec.js (12 tests — app shell boot)
Extended: core/inventory-crud.spec.js (53 tests — 7 domain blocks:
  required metal/type, seed guard, capsule fields, payment method,
  virtual sort options, serial CRUD journey, filter chip AND logic)

Source files removed after replacement assertions verified green.
Coverage map updated with consolidated decisions.
