--- tests/playwright/02-crud/crud.spec.js
@@ -1,599 +1,0 @@
-import { test, expect } from "@playwright/test";
-import { injectSeedInventory } from "../helpers/seed.js";
-import { installStakTrakrNetworkMocks } from "../helpers/mocks/routes.js";
-
-// Helper: count visible inventory cards (card style A/B/C)
-async function countCards(page) {
-  return page.locator("#cardViewGrid article").count();
-}
-
-// Helper: open the add-item modal
-async function openAddModal(page) {
-  // Dismiss any overlays that might intercept clicks (whatsNew popup, etc.)
-  const whatsNew = page.locator("#whatsNewPopup");
-  if (await whatsNew.isVisible()) {
-    await page.click("#whatsNewDismissBtn");
-  }
-  // Click via JS to bypass any overlay interception issues
-  await page.evaluate(() => document.getElementById("newItemBtn").click());
-  await expect(page.locator("#itemModal")).toBeVisible({ timeout: 10000 });
-}
-
-// Helper: submit the add-item form and wait for modal to close
-async function submitItemForm(page) {
-  await page.click("#itemModalSubmit");
-  await expect(page.locator("#itemModal")).toBeHidden();
-}
-
-// Helper: delete a BB-* item from table view by name, confirming the dialog
-async function deleteItemByName(page, name) {
-  const row = page.locator("#inventoryTable tbody tr").filter({ hasText: name });
-  await row
-    .locator(
-      'button[title*="elete"], button[aria-label*="elete"], button.delete-btn, td.actions button'
-    )
-    .first()
-    .click();
-  // Confirm the deletion dialog
-  const dialog = page
-    .locator('.modal:visible, [role="dialog"]:visible')
-    .filter({ hasText: /confirm|delete|remove/i });
-  if ((await dialog.count()) > 0) {
-    await dialog
-      .locator("button")
-      .filter({ hasText: /confirm|yes|delete|ok/i })
-      .first()
-      .click();
-  }
-}
-
-let sharedPage;
-
-test.describe.serial("02-crud", () => {
-  test.beforeAll(async ({ browser }) => {
-    sharedPage = await browser.newPage();
-    // Install shared network mocks before navigating to the app shell
-    await installStakTrakrNetworkMocks(sharedPage);
-    // Inject seed inventory data (includes ackVersion suppression for What's New popup)
-    await injectSeedInventory(sharedPage);
-    // Start in card view A so article elements are rendered for count assertions
-    await sharedPage.addInitScript(() => {
-      localStorage.setItem("cardViewStyle", "A");
-    });
-    await sharedPage.goto("/index.html", { waitUntil: "domcontentloaded" });
-    // Wait for app fully initialized: Add Item button visible + card grid rendered
-    await sharedPage.waitForSelector("#newItemBtn", { state: "visible" });
-    // Wait for JS init to complete — card grid must have rendered items
-    await sharedPage.waitForSelector("#cardViewGrid article", {
-      state: "attached",
-      timeout: 15000,
-    });
-    // Extra wait to ensure the deferred setupEventListeners() setTimeout has fired
-    await sharedPage.waitForTimeout(500);
-    // Dismiss whats-new popup if it still appears (race condition guard)
-    const popup = sharedPage.locator("#whatsNewPopup");
-    if (await popup.isVisible()) {
-      await sharedPage.click("#whatsNewDismissBtn");
-    }
-  });
-
-  test.afterAll(async () => {
-    if (!sharedPage) return;
-    // Cleanup: switch to table view and delete all BB-* items added by this suite
-    await sharedPage.click('[data-style="D"]');
-    await sharedPage.waitForSelector("#inventoryTable tbody tr", { state: "visible" });
-
-    const bbNames = [
-      "BB-SILVER-COIN",
-      "BB-GOLD-BAR",
-      "BB-GOLD-BAR-EDITED",
-      "BB-PLAT-ROUND",
-      "BB-PALL-BAR",
-      "BB-GOLDBACK",
-      "BB-OZ-ITEM",
-      "BB-G-ITEM",
-      "BB-KG-ITEM",
-      "BB-DWT-ITEM",
-      "BB-COUNT-TEST",
-    ];
-
-    for (const name of bbNames) {
-      const row = sharedPage.locator("#inventoryTable tbody tr").filter({ hasText: name });
-      if ((await row.count()) > 0) {
-        await deleteItemByName(sharedPage, name);
-        // Brief wait for re-render between deletes
-        await sharedPage.waitForTimeout(200);
-      }
-    }
-    await sharedPage.close();
-  });
-
-  test("2.1 — add item — Silver Coin", async () => {
-    // runbook: 02-crud.md §2.1
-    const page = sharedPage;
-    await openAddModal(page);
-    await page.selectOption("#itemMetal", "Silver");
-    await page.selectOption("#itemType", "Coin");
-    await page.fill("#itemName", "BB-SILVER-COIN");
-    await page.fill("#itemWeight", "1");
-    await page.fill("#itemPrice", "25");
-    await submitItemForm(page);
-
-    const count = await countCards(page);
-    expect(count).toBe(9);
-    await expect(page.locator("#cardViewGrid")).toContainText("BB-SILVER-COIN");
-  });
-
-  test("2.2 — add item — Gold Bar", async () => {
-    // runbook: 02-crud.md §2.2
-    const page = sharedPage;
-    await openAddModal(page);
-    await page.selectOption("#itemMetal", "Gold");
-    await page.selectOption("#itemType", "Bar");
-    await page.fill("#itemName", "BB-GOLD-BAR");
-    await page.fill("#itemWeight", "1");
-    await page.fill("#itemPrice", "2000");
-    await submitItemForm(page);
-
-    const count = await countCards(page);
-    expect(count).toBe(10);
-    await expect(page.locator("#cardViewGrid")).toContainText("BB-GOLD-BAR");
-  });
-
-  test("2.3 — add item — Platinum Round", async () => {
-    // runbook: 02-crud.md §2.3
-    const page = sharedPage;
-    await openAddModal(page);
-    await page.selectOption("#itemMetal", "Platinum");
-    await page.selectOption("#itemType", "Round");
-    await page.fill("#itemName", "BB-PLAT-ROUND");
-    await page.fill("#itemWeight", "1");
-    await page.fill("#itemPrice", "1000");
-    await submitItemForm(page);
-
-    const count = await countCards(page);
-    expect(count).toBe(11);
-    await expect(page.locator("#cardViewGrid")).toContainText("BB-PLAT-ROUND");
-  });
-
-  test("2.4 — add item — Palladium Bar", async () => {
-    // runbook: 02-crud.md §2.4
-    const page = sharedPage;
-    await openAddModal(page);
-    await page.selectOption("#itemMetal", "Palladium");
-    await page.selectOption("#itemType", "Bar");
-    await page.fill("#itemName", "BB-PALL-BAR");
-    await page.fill("#itemWeight", "1");
-    await page.fill("#itemPrice", "1000");
-    await submitItemForm(page);
-
-    const count = await countCards(page);
-    expect(count).toBe(12);
-    await expect(page.locator("#cardViewGrid")).toContainText("BB-PALL-BAR");
-  });
-
-  test("2.5 — add item — Goldback", async () => {
-    // runbook: 02-crud.md §2.5
-    // STAK-562 made Goldback a first-class Type; STAK-580 made Metal/Type required.
-    // Flow: pick Metal=Gold → Type=Goldback (auto-swaps weight unit to "gb" + denomination picker).
-    const page = sharedPage;
-    await openAddModal(page);
-    await page.selectOption("#itemMetal", "Gold");
-    await page.selectOption("#itemType", "Goldback");
-    await page.fill("#itemName", "BB-GOLDBACK");
-    // #itemGbDenom defaults to 1; denomination picker is shown, weight input is hidden
-    await page.fill("#itemPrice", "5");
-    await submitItemForm(page);
-
-    const count = await countCards(page);
-    expect(count).toBe(13);
-    await expect(page.locator("#cardViewGrid")).toContainText("BB-GOLDBACK");
-  });
-
-  test("2.6 — add items with each weight unit (oz, g, kg, lb)", async () => {
-    // runbook: 02-crud.md §2.6
-    // NOTE: The runbook lists oz/g/kg/dwt. "dwt" (pennyweight) is not in the
-    // #itemWeightUnit select (available options: oz, g, kg, lb, gb).
-    // BB-DWT-ITEM is added with "lb" as the closest non-default unit available;
-    // item name is preserved for cleanup purposes.
-    const page = sharedPage;
-
-    // BB-OZ-ITEM
-    await openAddModal(page);
-    await page.selectOption("#itemMetal", "Silver");
-    await page.selectOption("#itemType", "Coin");
-    await page.fill("#itemName", "BB-OZ-ITEM");
-    await page.selectOption("#itemWeightUnit", "oz");
-    await page.fill("#itemWeight", "1");
-    await page.fill("#itemPrice", "25");
-    await submitItemForm(page);
-
-    // BB-G-ITEM
-    await openAddModal(page);
-    await page.selectOption("#itemMetal", "Silver");
-    await page.selectOption("#itemType", "Coin");
-    await page.fill("#itemName", "BB-G-ITEM");
-    await page.selectOption("#itemWeightUnit", "g");
-    await page.fill("#itemWeight", "31.1");
-    await page.fill("#itemPrice", "25");
-    await submitItemForm(page);
-
-    // BB-KG-ITEM
-    await openAddModal(page);
-    await page.selectOption("#itemMetal", "Silver");
-    await page.selectOption("#itemType", "Coin");
-    await page.fill("#itemName", "BB-KG-ITEM");
-    await page.selectOption("#itemWeightUnit", "kg");
-    await page.fill("#itemWeight", "0.0311");
-    await page.fill("#itemPrice", "25");
-    await submitItemForm(page);
-
-    // BB-DWT-ITEM — using lb (dwt/pennyweight not available in form)
-    await openAddModal(page);
-    await page.selectOption("#itemMetal", "Silver");
-    await page.selectOption("#itemType", "Coin");
-    await page.fill("#itemName", "BB-DWT-ITEM");
-    await page.selectOption("#itemWeightUnit", "lb");
-    await page.fill("#itemWeight", "20");
-    await page.fill("#itemPrice", "25");
-    await submitItemForm(page);
-
-    const count = await countCards(page);
-    expect(count).toBe(17);
-    await expect(page.locator("#cardViewGrid")).toContainText("BB-OZ-ITEM");
-    await expect(page.locator("#cardViewGrid")).toContainText("BB-G-ITEM");
-    await expect(page.locator("#cardViewGrid")).toContainText("BB-KG-ITEM");
-    await expect(page.locator("#cardViewGrid")).toContainText("BB-DWT-ITEM");
-  });
-
-  test("2.7 — upload obverse image", async () => {
-    // runbook: 02-crud.md §2.7
-    const page = sharedPage;
-    // Open edit modal for BB-SILVER-COIN
-    const idx = await page.evaluate(() =>
-      window.inventory.findIndex((i) => i.name === "BB-SILVER-COIN")
-    );
-    expect(idx).toBeGreaterThanOrEqual(0);
-    await page.evaluate((i) => window.editItem(i), idx);
-    await expect(page.locator("#itemModal")).toBeVisible({ timeout: 10000 });
-
-    // Upload a 1x1 PNG via the hidden file input
-    const pngBuffer = Buffer.from(
-      "iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADUlEQVR42mNk+M9QDwADhgGAWjR9awAAAABJRU5ErkJggg==",
-      "base64"
-    );
-    await page.setInputFiles("#itemImageFileObv", {
-      name: "test-obv.png",
-      mimeType: "image/png",
-      buffer: pngBuffer,
-    });
-
-    // Wait for image processing and preview to appear
-    await expect(page.locator("#itemImagePreviewObv")).toBeVisible({ timeout: 10000 });
-    await submitItemForm(page);
-  });
-
-  test("2.8 — upload reverse image", async () => {
-    // runbook: 02-crud.md §2.8
-    const page = sharedPage;
-    const idx = await page.evaluate(() =>
-      window.inventory.findIndex((i) => i.name === "BB-SILVER-COIN")
-    );
-    expect(idx).toBeGreaterThanOrEqual(0);
-    await page.evaluate((i) => window.editItem(i), idx);
-    await expect(page.locator("#itemModal")).toBeVisible({ timeout: 10000 });
-
-    const pngBuffer = Buffer.from(
-      "iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADUlEQVR42mNk+M9QDwADhgGAWjR9awAAAABJRU5ErkJggg==",
-      "base64"
-    );
-    await page.setInputFiles("#itemImageFileRev", {
-      name: "test-rev.png",
-      mimeType: "image/png",
-      buffer: pngBuffer,
-    });
-
-    await expect(page.locator("#itemImagePreviewRev")).toBeVisible({ timeout: 10000 });
-    await submitItemForm(page);
-  });
-
-  test("2.9 — remove an uploaded image", async () => {
-    // runbook: 02-crud.md §2.9
-    const page = sharedPage;
-    const idx = await page.evaluate(() =>
-      window.inventory.findIndex((i) => i.name === "BB-SILVER-COIN")
-    );
-    expect(idx).toBeGreaterThanOrEqual(0);
-    await page.evaluate((i) => window.editItem(i), idx);
-    await expect(page.locator("#itemModal")).toBeVisible({ timeout: 10000 });
-
-    // The remove button for obverse should be visible if an image was previously uploaded
-    const removeBtn = page.locator("#itemImageRemoveBtnObv");
-    if (await removeBtn.isVisible()) {
-      await removeBtn.click();
-      // Preview should hide after removal
-      await expect(page.locator("#itemImagePreviewObv")).toBeHidden();
-    }
-    await submitItemForm(page);
-  });
-
-  test("2.10 — apply pattern-matched image (seed image rule)", async () => {
-    // runbook: 02-crud.md §2.10
-    // Verify the seed image / pattern-match APIs exist
-    const page = sharedPage;
-    const hasGetCustomRules = await page.evaluate(
-      () => typeof window.NumistaLookup?.getCustomRules === "function"
-    );
-    expect(hasGetCustomRules).toBe(true);
-
-    const hasImageCache = await page.evaluate(
-      () => typeof window.imageCache?.isAvailable === "function"
-    );
-    expect(hasImageCache).toBe(true);
-  });
-
-  test("2.11 — remove pattern-matched image", async () => {
-    // runbook: 02-crud.md §2.11
-    // Verify the image deletion API contract exists
-    const page = sharedPage;
-    const hasDeleteImages = await page.evaluate(
-      () => typeof window.imageCache?.deleteImages === "function"
-    );
-    expect(hasDeleteImages).toBe(true);
-
-    const hasDeleteMetadata = await page.evaluate(
-      () => typeof window.imageCache?.deleteMetadata === "function"
-    );
-    expect(hasDeleteMetadata).toBe(true);
-  });
-
-  test("2.12 — edit an existing item (all fields)", async () => {
-    // runbook: 02-crud.md §2.12
-    const page = sharedPage;
-    const idx = await page.evaluate(() =>
-      window.inventory.findIndex((i) => i.name === "BB-GOLD-BAR")
-    );
-    expect(idx).toBeGreaterThanOrEqual(0);
-
-    await page.evaluate((i) => window.editItem(i), idx);
-    await expect(page.locator("#itemModal")).toBeVisible({ timeout: 10000 });
-
-    // Verify modal is in edit mode
-    await expect(page.locator("#itemModalTitle")).toContainText("Edit");
-
-    // Change name and price
-    await page.fill("#itemName", "BB-GOLD-BAR-EDITED");
-    await page.fill("#itemPrice", "2100");
-    await submitItemForm(page);
-
-    // Verify the grid reflects the edit
-    await expect(page.locator("#cardViewGrid")).toContainText("BB-GOLD-BAR-EDITED");
-    const count = await countCards(page);
-    expect(count).toBe(17);
-  });
-
-  test("2.13 — delete item — confirmation dialog appears before delete", async () => {
-    // runbook: 02-crud.md §2.13
-    const page = sharedPage;
-    const idx = await page.evaluate(() =>
-      window.inventory.findIndex((i) => i.name === "BB-PALL-BAR")
-    );
-    expect(idx).toBeGreaterThanOrEqual(0);
-
-    await page.evaluate((i) => window.deleteItem(i), idx);
-
-    // Verify the remove modal is visible with the item name
-    await expect(page.locator("#removeItemModal")).toBeVisible({ timeout: 5000 });
-    await expect(page.locator("#removeItemName")).toContainText("BB-PALL-BAR");
-
-    // Confirm deletion
-    await page.click("#removeItemDeleteBtn");
-
-    // Verify modal closes and count decreases
-    await expect(page.locator("#removeItemModal")).toBeHidden();
-    const count = await countCards(page);
-    expect(count).toBe(16);
-  });
-
-  test("2.14 — item count badge updates after add/edit/delete", async () => {
-    // runbook: 02-crud.md §2.14
-    const page = sharedPage;
-
-    // Read current badge count (should be 16 after 2.13 deleted BB-PALL-BAR)
-    const initialCount = await page.locator("#totalItemsAll").textContent();
-    expect(parseInt(initialCount, 10)).toBe(16);
-
-    // Add a new item
-    await openAddModal(page);
-    await page.selectOption("#itemMetal", "Silver");
-    await page.selectOption("#itemType", "Coin");
-    await page.fill("#itemName", "BB-COUNT-TEST");
-    await page.fill("#itemWeight", "1");
-    await page.fill("#itemPrice", "25");
-    await submitItemForm(page);
-
-    // Count should increase by 1
-    const afterAdd = await page.locator("#totalItemsAll").textContent();
-    expect(parseInt(afterAdd, 10)).toBe(17);
-
-    // Delete BB-COUNT-TEST
-    const idx = await page.evaluate(() =>
-      window.inventory.findIndex((i) => i.name === "BB-COUNT-TEST")
-    );
-    expect(idx).toBeGreaterThanOrEqual(0);
-    await page.evaluate((i) => window.deleteItem(i), idx);
-    await expect(page.locator("#removeItemModal")).toBeVisible({ timeout: 5000 });
-    await page.click("#removeItemDeleteBtn");
-    await expect(page.locator("#removeItemModal")).toBeHidden();
-
-    // Count should return to original
-    const afterDelete = await page.locator("#totalItemsAll").textContent();
-    expect(parseInt(afterDelete, 10)).toBe(16);
-  });
-
-  test("2.15 — search for an item by name", async () => {
-    // runbook: 02-crud.md §2.15
-    const page = sharedPage;
-
-    // Type search query — fill dispatches an input event which triggers the debounced search
-    await page.fill("#searchInput", "BB-SILVER");
-    // Wait for debounce (300ms) + render
-    await page.waitForTimeout(500);
-
-    // BB-SILVER-COIN should appear; BB-GOLD-BAR-EDITED should not
-    await expect(page.locator("#cardViewGrid")).toContainText("BB-SILVER-COIN");
-    const grid = page.locator("#cardViewGrid");
-    await expect(grid).not.toContainText("BB-GOLD-BAR-EDITED");
-  });
-
-  test("2.16 — filter chips reflect search results", async () => {
-    // runbook: 02-crud.md §2.16
-    const page = sharedPage;
-
-    // Search should still be active from 2.15 — chips render in #activeFilters
-    const chipContainer = page.locator("#activeFilters");
-    await expect(chipContainer).toBeVisible();
-    // There should be at least one filter chip rendered (search chip + category chips)
-    const chipCount = await chipContainer.locator(".filter-chip").count();
-    expect(chipCount).toBeGreaterThan(0);
-
-    // Clear search to reset for subsequent tests
-    await page.evaluate(() => window.clearAllFilters());
-    await page.waitForTimeout(300);
-  });
-
-  test("2.17 — filter via chip (metal)", async () => {
-    // runbook: 02-crud.md §2.17
-    const page = sharedPage;
-
-    // Ensure filters are clear
-    await page.evaluate(() => window.clearAllFilters());
-    await page.waitForTimeout(300);
-
-    const beforeCount = await countCards(page);
-    expect(beforeCount).toBe(16);
-
-    // Click a "Silver" filter chip
-    const silverChip = page.locator(".filter-chip").filter({ hasText: "Silver" }).first();
-    await expect(silverChip).toBeVisible({ timeout: 5000 });
-    await silverChip.click();
-    await page.waitForTimeout(300);
-
-    // Only silver items should appear — count should be less than total
-    const afterCount = await countCards(page);
-    expect(afterCount).toBeGreaterThan(0);
-    expect(afterCount).toBeLessThan(beforeCount);
-  });
-
-  test("2.18 — remove filter chip narrows/expands results", async () => {
-    // runbook: 02-crud.md §2.18
-    const page = sharedPage;
-
-    // Silver filter should still be active from 2.17
-    const filteredCount = await countCards(page);
-    expect(filteredCount).toBeLessThan(16);
-
-    // Click the active Silver chip to deactivate it
-    const activeChip = page
-      .locator(".filter-chip.filter-chip-active")
-      .filter({ hasText: "Silver" })
-      .first();
-    await activeChip.click();
-    await page.waitForTimeout(300);
-
-    // All items should reappear
-    const afterCount = await countCards(page);
-    expect(afterCount).toBe(16);
-
-    // Clear all filters to reset state for next tests
-    await page.evaluate(() => window.clearAllFilters());
-    await page.waitForTimeout(300);
-  });
-
-  test("2.19 — switch card views A → B → C → Table (D)", async () => {
-    // runbook: 02-crud.md §2.19
-    const page = sharedPage;
-
-    // Style A — card grid with articles
-    await page.click('[data-style="A"]');
-    await page.waitForTimeout(300);
-    const cardsA = await countCards(page);
-    expect(cardsA).toBe(16);
-
-    // Style B — different card layout, still articles
-    await page.click('[data-style="B"]');
-    await page.waitForTimeout(300);
-    const cardsB = await countCards(page);
-    expect(cardsB).toBe(16);
-
-    // Style C — another card layout
-    await page.click('[data-style="C"]');
-    await page.waitForTimeout(300);
-    const cardsC = await countCards(page);
-    expect(cardsC).toBe(16);
-
-    // Style D — table view
-    await page.click('[data-style="D"]');
-    await page.waitForTimeout(300);
-    await expect(page.locator("#inventoryTable")).toBeVisible();
-    const tableRows = await page.locator("#inventoryTable tbody tr").count();
-    expect(tableRows).toBeGreaterThan(0);
-
-    // Switch back to A for subsequent tests
-    await page.click('[data-style="A"]');
-    await page.waitForTimeout(300);
-    const cardsAfter = await countCards(page);
-    expect(cardsAfter).toBe(16);
-  });
-
-  test("2.20 — melt value recalculates correctly when spot price changes", async () => {
-    // runbook: 02-crud.md §2.20
-    const page = sharedPage;
-
-    // Get the current melt value for BB-SILVER-COIN by computing it directly
-    const initialMelt = await page.evaluate(() => {
-      const item = window.inventory.find((i) => i.name === "BB-SILVER-COIN");
-      if (!item) return null;
-      const spot = spotPrices.silver || 0;
-      return typeof computeMeltValue === "function" ? computeMeltValue(item, spot) : null;
-    });
-
-    // Override the silver spot price to a known value
-    await page.evaluate(() => {
-      spotPrices.silver = 50;
-      if (typeof renderTable === "function") renderTable();
-    });
-    await page.waitForTimeout(300);
-
-    // Recalculate melt value with the new spot price
-    const updatedMelt = await page.evaluate(() => {
-      const item = window.inventory.find((i) => i.name === "BB-SILVER-COIN");
-      if (!item) return null;
-      return typeof computeMeltValue === "function" ? computeMeltValue(item, 50) : null;
-    });
-
-    // Melt value should have changed (new spot = $50)
-    expect(updatedMelt).not.toBeNull();
-    expect(updatedMelt).toBeGreaterThan(0);
-    if (initialMelt !== null && initialMelt > 0) {
-      expect(updatedMelt).not.toBe(initialMelt);
-    }
-  });
-
-  test("2.21 — cleanup — delete all BB-* test items", async () => {
-    // runbook: 02-crud.md §2.21
-    // Verify BB-* items exist in inventory before afterAll cleans them up
-    const page = sharedPage;
-    const bbItems = await page.evaluate(() =>
-      window.inventory.filter((i) => i.name && i.name.startsWith("BB-")).map((i) => i.name)
-    );
-
-    // We should still have BB-* items from tests 2.1–2.6 and 2.12
-    // (BB-PALL-BAR was deleted in 2.13, BB-COUNT-TEST was deleted in 2.14)
-    expect(bbItems.length).toBeGreaterThan(0);
-
-    // Verify the delete mechanism works by confirming deleteItem is callable
-    const hasDeleteItem = await page.evaluate(() => typeof window.deleteItem === "function");
-    expect(hasDeleteItem).toBe(true);
-  });
-});
